Initial commit of the 'gpuallocator' package
The `gpuallocator` package provides a generic abstraction for performing
GPU allocations independent of the larger system the `gpuallocator` is
integrated with.

The abstractions provided by this package are not meant to do actual
"allocation" of GPUs to any specific entity, but rather run the
algorithm responsible for deciding which GPUs should be chosen for
allocation based on the set of GPUs available in the system and the
number of GPUs being requested.

Different policies can be hooked in to run different allocation
algorithms depending on the specific needs of the system.

For example, a system like Kubernetes would use this package to help
decide which subset of GPUs to hand out to a container once it has
decided how many GPUs it should be granted. The policy it would choose
would be based on the topological ordering of GPUs to ensure optimal
affinity for groups of GPUs allocated to the same container.

The primary object provided by this package is the `Allocator` object,
and the primary interface used to decide how allocation should actually
occur is called `Policy`.

Signed-off-by: Kevin Klues <kklues@nvidia.com>