Fix construction of linked devices
This change fixes the construction of the linked devices.

Firstly, the nvmlDevices are used directly when determining link
properties instead of relying on the local Device type. This allows
the nvml Device handle to be used directly.

In addition an ERROR_INVALID_ARGUMENT when querying link state is
treated as a non-fatal error to handle cases where older drivers are
used. This also aligns with the former implementation in gpu-monitoring-tools.

Signed-off-by: Evan Lezar <elezar@nvidia.com>